requirements.txt
# Astronomer Airflow Requirements
# Note: apache-airflow is provided by the Astronomer base image

# Core dependencies for YouTube pipeline
python-dotenv>=1.0.0
yt-dlp>=2023.11.16
openai>=1.3.0
google-api-python-client>=2.108.0  # YouTube Data API v3

# Vector Database - updated to newer version with better wheel support
pymilvus>=2.4.0
grpcio-tools  # Pre-built wheels available

# Utilities
httpx==0.25.2
mmh3==3.0.0
requests>=2.31.0
pydantic>=2.5.0

# Optional: Web Framework (only needed if running API in Airflow)
# fastapi==0.104.1
# uvicorn[standard]==0.24.0
# jinja2==3.1.2
# python-multipart==0.0.6
